+++ src/ra/map.cc
@@ -1617,7 +1617,7 @@ int MapClass::Validate() {
       tclass = &TemplateTypeClass::As_Reference(ttype);
       ticon = (*this)[cell].TIcon;
       Mem_Copy(Get_Icon_Set_Map(tclass->Get_Image_Data()), map,
-               tclass->Width * tclass->Height);
+               static_cast<unsigned long>(tclass->Width) * tclass->Height);
       if (ticon < 0 || ticon >= tclass->Width * tclass->Height ||
           map[ticon] == 0xff) {
         return false;
