+++ src/td/gadget.cc
@@ -457,7 +457,7 @@ KeyNumType GadgetClass::Input() {
       */
       GraphicBufferClass temp_page(
           SeenBuff.Get_Width(), SeenBuff.Get_Height(), nullptr,
-          SeenBuff.Get_Width() * SeenBuff.Get_Height());
+          static_cast<long>(SeenBuff.Get_Width()) * SeenBuff.Get_Height());
       char filename[30];
 
       SeenBuff.Blit(temp_page);
