+++ src/td/heap.cc
@@ -76,6 +76,7 @@
 #include "td/unit.h"
 #include "tech/noinit.h"
 #include "tech/wwfile.h"
+#include <cstddef>
 
 /***********************************************************************************************
  * FixedHeapClass::FixedHeapClass -- Normal constructor for heap management
@@ -167,7 +168,7 @@ int FixedHeapClass::Set_Heap(int count, void* buffer) {
   */
   FreeFlag.resize(count, false);
   if (!buffer) {
-    buffer = new char[count * Size];
+    buffer = new char[static_cast<std::size_t>(count) * Size];
     if (!buffer) {
       FreeFlag.clear();
       return false;
