+++ src/td/interpal.cc
@@ -75,7 +75,7 @@ void Read_Interpolation_Palette(const char* palette_file_name) {
 
   if (palette_file.Is_Available()) {
     palette_file.Open(FileAccess::kRead);
-    palette_file.Read(&PaletteInterpolationTable[0][0], 256 * 256);
+    palette_file.Read(&PaletteInterpolationTable[0][0], static_cast<long>(256) * 256);
     palette_file.Close();
     InterpolationPaletteChanged = false;
   }
@@ -101,7 +101,7 @@ void Write_Interpolation_Palette(const char* palette_file_name) {
 
   if (!palette_file.Is_Available()) {
     palette_file.Open(FileAccess::kWrite);
-    palette_file.Write(&PaletteInterpolationTable[0][0], 256 * 256);
+    palette_file.Write(&PaletteInterpolationTable[0][0], static_cast<long>(256) * 256);
     palette_file.Close();
   }
 }
